aimhi-chatbot/requirements.txt
Flask==2.3.0
Flask-Limiter>=2.3
Flask-CORS>=4.0.0
transitions==0.9.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
supabase>=2.0.0
PyJWT>=2.8.0
